ft_memcmp.c

+++ ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: bsibanyo <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/06/05 09:59:34 by bsibanyo          #+#    #+#             */
-/*   Updated: 2019/06/05 14:30:52 by bsibanyo         ###   ########.fr       */
+/*   Updated: 2019/06/05 15:28:25 by bsibanyo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ int	ft_memcmp(const void *s1, const void *s2, size_t n)
 	{
 		if (str1[i] != str2[i])
 		{
-			return (str1[i] - str2[i]);
+			return ((int)str1[i] - (int)str2[i]);
 			i++;
 		}
 	}
